backend/requirements.txt
# Core Framework
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# LangChain & LangGraph
langgraph==0.2.28
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0

# Database
pymongo==4.6.1
psycopg2-binary==2.9.9
pgvector==0.2.4

# Document Processing
pypdf==4.0.0
python-pptx==0.6.23
python-docx==1.1.0
openpyxl==3.1.2
pillow==10.2.0
pdf2image==1.17.0

# Vector & Search
numpy==1.26.3
scikit-learn==1.4.0
sentence-transformers==2.3.1

# Utilities
pydantic==2.5.3
tenacity==8.2.3
tiktoken==0.5.2
requests==2.31.0

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
faker==22.6.0
